refactoring com if to use separate thread
